move keras dependencies into separate LayerPacker class outside of _shared
